phising-api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-identity==1.21.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
configparser==7.2.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
googlesearch-python==1.3.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.66.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pysafebrowsing==0.1.3
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
tldextract==5.1.3
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
